+++ tests/yaml/test_match_to_pgml.py
@@ -1,4 +1,3 @@
-#!/usr/bin/env python3
 
 import lib_test_utils